Package name 'gr.ekt.repositories.dspace.utils' removed. All functions in gr.ekt.repositories.dspace.utils.Utilities.java were transfered in org.dspace.app.util.Util.java Documentation added for the above functions
